feat: renamed all versioned plugin components with version suffix
