+++ app/src/pages/staff-pages/settings/components/extensions/extensions.tsx
@@ -1,16 +1,21 @@
 import React, {useEffect, useState} from 'react';
-import {Box, Card, CardContent, CardHeader, Grid, Typography} from '@mui/material';
-import {ServiceProviderApiService} from '../../../../../services/service-provider-api-service';
-import {ServiceProviderDTO} from '../../../../../models/dtos/service-provider-dto';
+import {Box, Card, CardContent, CardHeader, Grid, IconButton, Tooltip, Typography} from '@mui/material';
 import {
     PluginComponentTypeDisplayNames,
     PluginComponentTypeOptions,
     PluginDTO,
     PluginsApiService,
 } from '../../../../../services/plugins-api-service';
+import {AlertComponent} from '../../../../../components/alert/alert-component';
+import ReactMarkdown from 'react-markdown';
+import {Badge} from '../../../../../components/badge/badge';
+import Article from '@aivot/mui-material-symbols-400-outlined/dist/article/Article';
+import {useConfirm} from '../../../../../providers/confirm-provider';
 
 export function Extensions() {
     const [plugins, setPlugins] = useState<PluginDTO[]>([]);
+
+    const confirm = useConfirm();
 
     useEffect(() => {
         new PluginsApiService()
@@ -52,22 +57,91 @@ export function Extensions() {
                             }}
                         >
                             <CardHeader
-                                title={serviceProvider.name + ' (v' + serviceProvider.version + ')'}
-                                titleTypographyProps={{variant: 'h6'}}
-                                subheaderTypographyProps={{
-                                    variant: 'body2',
-                                    color: 'text.secondary',
-                                    fontSize: '0.875rem',
+                                title={serviceProvider.name}
+                                subheader={
+                                    <Box
+                                        component="ul"
+                                        sx={{
+                                            listStyle: 'none',
+                                            m: 0,
+                                            p: 0,
+                                            display: 'flex',
+                                            '& li::before': {
+                                                content: '"•"',
+                                                color: 'text.secondary',
+                                                display: 'inline-block',
+                                                mx: 0.5,
+                                            },
+                                            '& li:first-child::before': {
+                                                display: 'none',
+                                            },
+                                        }}
+                                    >
+                                        <li>Version {serviceProvider.version}</li>
+                                        <li>
+                                            {serviceProvider.vendorName}&nbsp;
+                                            <a href={serviceProvider.vendorWebsite}>{serviceProvider.vendorWebsite}</a>
+                                        </li>
+                                    </Box>
+                                }
+                                slotProps={{
+                                    title: {
+                                        variant: 'h6',
+                                    },
+                                    subheader: {
+                                        variant: 'body2',
+                                        color: 'text.secondary',
+                                        fontSize: '0.875rem',
+                                    },
                                 }}
+                                action={
+                                    <Tooltip title="Changelog anzeigen">
+                                        <IconButton
+                                            onClick={() => {
+                                                confirm({
+                                                    title: 'Changelog',
+                                                    children:
+                                                        <ReactMarkdown>{serviceProvider.changelog}</ReactMarkdown>,
+                                                    hideCancelButton: true,
+                                                    confirmButtonText: 'Schließen',
+                                                });
+                                            }}
+                                        >
+                                            <Article/>
+                                        </IconButton>
+                                    </Tooltip>
+                                }
                             />
-                            <CardContent sx={{flexGrow: 1}}>
-                                <Typography
-                                    variant="body2"
-                                    color="text.secondary"
-                                >
-                                    {serviceProvider.description}
-                                </Typography>
 
+                            <CardContent
+                                sx={{
+                                    flexGrow: 1,
+                                }}
+                            >
+                                <Box
+                                    sx={{
+                                        color: 'text.secondary',
+                                    }}
+                                >
+                                    <ReactMarkdown>
+                                        {serviceProvider.description}
+                                    </ReactMarkdown>
+                                </Box>
+
+                                {
+                                    serviceProvider.deprecationNotice != null &&
+                                    <AlertComponent
+                                        title="Veraltete Erweiterung"
+                                        color="warning"
+                                        sx={{
+                                            mt: 2,
+                                        }}
+                                    >
+                                        <ReactMarkdown>
+                                            {serviceProvider.deprecationNotice}
+                                        </ReactMarkdown>
+                                    </AlertComponent>
+                                }
 
                                 {
                                     PluginComponentTypeOptions.map((componentType) => (
@@ -90,15 +164,98 @@ export function Extensions() {
                                                         </li>
                                                     )
                                                 }
+
                                                 {
                                                     serviceProvider
                                                         .components
                                                         .filter(group => group[0].componentType === componentType)
-                                                        .map((componentGroup) => (
-                                                            <li key={componentGroup[0].name}>
-                                                                <strong>{componentGroup[0].name} (v{componentGroup[0].componentVersion})</strong>: {componentGroup[0].description}
-                                                            </li>
-                                                        ))
+                                                        .map((componentGroup) => {
+                                                            const orderedVersions = componentGroup.sort((a, b) => b.majorVersion - a.majorVersion);
+                                                            const currentVersion = orderedVersions[0];
+                                                            return (
+                                                                <li key={currentVersion.name}>
+                                                                    <Box>
+                                                                        <Box
+                                                                            sx={{
+                                                                                display: 'flex',
+                                                                                alignItems: 'center',
+                                                                                gap: 1,
+                                                                            }}
+                                                                        >
+                                                                            {currentVersion.name}
+                                                                            <Badge label={currentVersion.componentVersion}
+                                                                                   color="default"/>
+                                                                        </Box>
+
+                                                                        <Typography
+                                                                            color="text.secondary"
+                                                                        >
+                                                                            <ReactMarkdown>
+                                                                                {currentVersion.description}
+                                                                            </ReactMarkdown>
+                                                                        </Typography>
+
+                                                                        {
+                                                                            currentVersion.deprecationNotice != null &&
+                                                                            <AlertComponent
+                                                                                title="Veraltete Komponente"
+                                                                                color="warning"
+                                                                                sx={{
+                                                                                    mt: 1,
+                                                                                }}
+                                                                            >
+                                                                                <ReactMarkdown>
+                                                                                    {currentVersion.deprecationNotice}
+                                                                                </ReactMarkdown>
+                                                                            </AlertComponent>
+                                                                        }
+
+                                                                        {
+                                                                            orderedVersions.length > 1 &&
+                                                                            orderedVersions.slice(1).map((version) => (
+                                                                                <Box
+                                                                                    key={version.name}
+                                                                                    sx={{
+                                                                                        mt: 1,
+                                                                                        ml: 2,
+                                                                                        borderLeft: '2px solid',
+                                                                                        borderColor: 'divider',
+                                                                                        pl: 2,
+                                                                                    }}
+                                                                                >
+                                                                                    <Box
+                                                                                        sx={{
+                                                                                            display: 'flex',
+                                                                                            alignItems: 'center',
+                                                                                            gap: 1,
+                                                                                        }}
+                                                                                    >
+                                                                                        {version.name}
+                                                                                        <Badge label={version.componentVersion}
+                                                                                               color="default"/>
+                                                                                    </Box>
+
+                                                                                    {
+                                                                                        version.deprecationNotice != null &&
+                                                                                        <AlertComponent
+                                                                                            title="Veraltete Komponente"
+                                                                                            color="warning"
+                                                                                            sx={{
+                                                                                                mt: 1,
+                                                                                            }}
+                                                                                        >
+                                                                                            <ReactMarkdown>
+                                                                                                {version.deprecationNotice}
+                                                                                            </ReactMarkdown>
+                                                                                        </AlertComponent>
+                                                                                    }
+                                                                                </Box>
+                                                                            ))
+                                                                        }
+                                                                    </Box>
+                                                                </li>
+                                                            );
+                                                        })
                                                 }
                                             </ul>
                                         </Box>
