+++ src/main/java/de/aivot/GoverBackend/plugins/core/v1/javascript/DataObjectJavascriptV1.java
@@ -9,7 +9,6 @@ import de.aivot.GoverBackend.dataObject.services.DataObjectSchemaService;
 import de.aivot.GoverBackend.javascript.providers.JavascriptFunctionProvider;
 import de.aivot.GoverBackend.javascript.services.JavascriptEngine;
 import de.aivot.GoverBackend.lib.exceptions.ResponseException;
-import de.aivot.GoverBackend.plugin.models.PluginComponent;
 import de.aivot.GoverBackend.plugins.core.Core;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
@@ -31,13 +30,13 @@ import java.util.Map;
  * The functions are exposed to the JavaScript environment through the GraalVM Polyglot API.
  */
 @Component
-public class DataObjectJavascript implements JavascriptFunctionProvider {
+public class DataObjectJavascriptV1 implements JavascriptFunctionProvider {
     private final DataObjectSchemaService dataObjectSchemaService;
     private final DataObjectItemService dataObjectItemService;
 
     @Autowired
-    public DataObjectJavascript(DataObjectSchemaService dataObjectSchemaService,
-                                @Lazy DataObjectItemService dataObjectItemService) {
+    public DataObjectJavascriptV1(DataObjectSchemaService dataObjectSchemaService,
+                                  @Lazy DataObjectItemService dataObjectItemService) {
         this.dataObjectSchemaService = dataObjectSchemaService;
         this.dataObjectItemService = dataObjectItemService;
     }
@@ -140,7 +139,7 @@ public class DataObjectJavascript implements JavascriptFunctionProvider {
 
         var items = page
                 .stream()
-                .map(DataObjectJavascript::getItemData)
+                .map(DataObjectJavascriptV1::getItemData)
                 .toList();
 
         return JavascriptEngine
