+++ src/main/java/de/aivot/GoverBackend/plugins/core/v1/nodes/actions/LowCodeActionNodeV1.java
@@ -5,34 +5,26 @@ import de.aivot.GoverBackend.elements.models.elements.form.input.CodeInputElemen
 import de.aivot.GoverBackend.elements.models.elements.layout.ConfigLayoutElement;
 import de.aivot.GoverBackend.javascript.models.JavascriptCode;
 import de.aivot.GoverBackend.javascript.services.JavascriptEngineFactoryService;
-import de.aivot.GoverBackend.plugin.models.PluginComponent;
 import de.aivot.GoverBackend.plugins.core.Core;
-import de.aivot.GoverBackend.process.entities.ProcessEntity;
-import de.aivot.GoverBackend.process.entities.ProcessNodeEntity;
-import de.aivot.GoverBackend.process.entities.ProcessVersionEntity;
-import de.aivot.GoverBackend.process.entities.ProcessInstanceEntity;
-import de.aivot.GoverBackend.process.enums.ProcessHistoryEventType;
 import de.aivot.GoverBackend.process.enums.ProcessNodeType;
 import de.aivot.GoverBackend.process.exceptions.ProcessNodeExecutionException;
 import de.aivot.GoverBackend.process.exceptions.ProcessNodeExecutionExceptionUnknown;
 import de.aivot.GoverBackend.process.models.*;
 import de.aivot.GoverBackend.process.services.ProcessDataService;
-import de.aivot.GoverBackend.user.entities.UserEntity;
 import jakarta.annotation.Nonnull;
-import jakarta.annotation.Nullable;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
 import java.util.Map;
 
 @Component
-public class LowCodeActionNode implements ProcessNodeDefinition {
+public class LowCodeActionNodeV1 implements ProcessNodeDefinition {
     private static final String PORT_NAME = "output";
 
     private static final String CODE_FIELD_KEY = "js_code";
     private final JavascriptEngineFactoryService javascriptEngineFactoryService;
 
-    public LowCodeActionNode(JavascriptEngineFactoryService javascriptEngineFactoryService) {
+    public LowCodeActionNodeV1(JavascriptEngineFactoryService javascriptEngineFactoryService) {
         this.javascriptEngineFactoryService = javascriptEngineFactoryService;
     }
