+++ src/main/java/de/aivot/GoverBackend/plugins/core/v1/storage/LocalDiskStorageProviderDefinitionV1.java
@@ -8,7 +8,6 @@ import de.aivot.GoverBackend.elements.models.elements.layout.ConfigLayoutElement
 import de.aivot.GoverBackend.elements.utils.ElementPOJOMapper;
 import de.aivot.GoverBackend.enums.ElementType;
 import de.aivot.GoverBackend.lib.exceptions.ResponseException;
-import de.aivot.GoverBackend.plugin.models.PluginComponent;
 import de.aivot.GoverBackend.plugins.core.Core;
 import de.aivot.GoverBackend.storage.exceptions.StorageException;
 import de.aivot.GoverBackend.storage.models.StorageDocument;
@@ -27,7 +26,7 @@ import java.util.LinkedList;
 import java.util.Optional;
 
 @Component
-public class LocalDiskStorageProviderDefinition implements StorageProviderDefinition<LocalDiskStorageProviderDefinition.Config> {
+public class LocalDiskStorageProviderDefinitionV1 implements StorageProviderDefinition<LocalDiskStorageProviderDefinitionV1.Config> {
     @Nonnull
     @Override
     public String getParentPluginKey() {
@@ -69,7 +68,7 @@ public class LocalDiskStorageProviderDefinition implements StorageProviderDefini
     }
 
     @Override
-    public Class<LocalDiskStorageProviderDefinition.Config> getConfigClass() {
+    public Class<LocalDiskStorageProviderDefinitionV1.Config> getConfigClass() {
         return Config.class;
     }
